refactor(*):unifying styles in a single directory
