+++ packages/geo/src/lib/datasource/utils/index.ts
@@ -1,2 +1,1 @@
-export * from './esri-style-generator';
 export * from './tilegrid';
