+++ packages/geo/src/lib/geo.module.ts
@@ -20,6 +20,7 @@ import { IgoSearchModule } from './search/search.module';
 import { IgoToastModule } from './toast/toast.module';
 import { IgoGeoWorkspaceModule } from './workspace/workspace.module';
 import { IgoWktModule } from './wkt/wkt.module';
+import { IgoStyleModule } from './style/style.module';
 
 @NgModule({
   imports: [],
@@ -44,6 +45,7 @@ import { IgoWktModule } from './wkt/wkt.module';
     IgoSearchModule,
     IgoToastModule,
     IgoGeoWorkspaceModule,
+    IgoStyleModule,
     IgoWktModule
   ]
 })
