+++ packages/geo/src/lib/import-export/index.ts
@@ -1,4 +1,3 @@
 export * from './shared';
 export * from './import-export';
 export * from './export-button';
-export * from './style-list';
