+++ packages/geo/src/lib/import-export/shared/import.utils.ts
@@ -15,9 +15,9 @@ import {
 import { VectorLayer } from '../../layer/shared/layers/vector-layer';
 import { IgoMap } from '../../map/shared/map';
 import { QueryableDataSourceOptions } from '../../query/shared/query.interfaces';
-import { StyleService } from '../../layer/shared/style.service';
-import { StyleByAttribute } from '../../layer/shared/vector-style.interface';
-import { StyleListService } from '../style-list/style-list.service';
+import { StyleService } from '../../style/style-service/style.service';
+import { StyleByAttribute } from '../../style/shared/vector/vector-style.interface';
+import { StyleListService } from '../../style/style-list/style-list.service';
 import { ClusterParam } from '../../layer/shared/clusterParam';
 import { ClusterDataSource } from '../../datasource/shared/datasources/cluster-datasource';
 import { ClusterDataSourceOptions } from '../../datasource/shared/datasources/cluster-datasource.interface';
