+++ packages/geo/src/lib/layer/layer.module.ts
@@ -28,7 +28,7 @@ import {
 } from '@igo2/common';
 
 import { LayerService } from './shared/layer.service';
-import { StyleService } from './shared/style.service';
+import { StyleService } from '../style/style-service/style.service';
 import { LayerListToolService } from './layer-list-tool/layer-list-tool.service';
 import { LayerItemComponent } from './layer-item/layer-item.component';
 import { LayerLegendComponent } from './layer-legend/layer-legend.component';
