+++ packages/geo/src/lib/layer/shared/index.ts
@@ -1,6 +1,4 @@
 export * from './layer.service';
 export * from './layer.enums';
 export * from './layers';
-export * from './style.service';
 export * from './clusterParam';
-export * from './vector-style.interface';
