+++ packages/geo/src/lib/overlay/shared/index.ts
@@ -3,4 +3,3 @@ export * from './overlay.directive';
 export * from './overlay.service';
 export * from './overlay.enum';
 export * from './overlay.utils';
-export * from './overlay-marker-style.utils';
