+++ packages/geo/src/lib/style/shared/datasource/esri-style-generator.ts
