+++ packages/geo/src/lib/style/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './datasource';
+export * from './vector';
+export * from './overlay';
