+++ packages/geo/src/lib/style/shared/overlay/index.ts
@@ -1,0 +1,2 @@
+export * from './overlay-marker-style.utils';
+export * from './overlay-style.utils';
