+++ packages/geo/src/lib/style/shared/overlay/overlay-marker-style.utils.ts
