+++ packages/geo/src/lib/style/shared/vector/index.ts
@@ -1,0 +1,2 @@
+export * from './commonVectorStyle'
+export * from './vector-style.interface'
