+++ packages/geo/src/lib/style/style-service/index.ts
@@ -1,0 +1,1 @@
+export * from './style.service';
