--- packages/geo/src/lib/utils/commonVectorStyle.interface.ts
@@ -1,18 +1,0 @@
-import olFeature from 'ol/Feature';
-import type { default as OlGeometry } from 'ol/geom/Geometry';
-import { Feature } from '../feature/shared/feature.interfaces';
-
-export interface FeatureCommonVectorStyleOptions extends CommonVectorStyleOptions {
-  feature: Feature | olFeature<OlGeometry>;
-}
-
-export interface CommonVectorStyleOptions {
-  markerColor?: string | number[]; // marker fill
-  markerOpacity?: number; // marker opacity, not applied if a rgba is provided
-  markerOutlineColor?: string | number[]; // marker contour
-  fillColor?: string | number[]; // poly
-  fillOpacity?: number; // poly fill opacity, not applied if a rgba is provided
-  strokeColor?: string | number[]; // line and poly
-  strokeOpacity?: number; // line and poly, not applied if a rgba is provided
-  strokeWidth?: number; // line and poly
-}
