+++ packages/geo/src/lib/utils/index.ts
@@ -1,5 +1,3 @@
 export * from './googleLinks';
 export * from './id-generator';
-export * from './commonVectorStyle';
-export * from './commonVectorStyle.interface';
 export * from './osmLinks';
